Fix tier adjustment on actions that use standard attack damage
