+++ .gitattributes
@@ -1,0 +1,2 @@
+* text=auto eol=lf
+*.json text eol=lf
